move downloadlink to components
